+++ src/main.c
@@ -363,7 +363,7 @@ struct Configurable {
 };
 
 #define WARN_PREFIX "Warning: "
-#define WARN_TEXTWIDTH (79 - strlen(WARN_PREFIX))
+#define WARN_TEXTWIDTH (79 - (int)strlen(WARN_PREFIX))
 /* produce this text message to the user unless mute was selected */
 static void warnf(struct Configurable *config, const char *fmt, ...)
 {
